Loi layout

+++ src/main/java/com/shopthoitrangnike/shopThoiTrangNike/service/CategoryService.java
@@ -1,31 +1,64 @@
 package com.shopthoitrangnike.shopThoiTrangNike.service;
 
-
 import com.shopthoitrangnike.shopThoiTrangNike.model.Category;
 import com.shopthoitrangnike.shopThoiTrangNike.repository.CategoryRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import jakarta.validation.constraints.NotNull;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
+/**
+ * Service class for managing categories.
+ */
 @Service
+@RequiredArgsConstructor
+@Transactional
 public class CategoryService {
-    @Autowired
-    private CategoryRepository categoryRepository;
-
-    public List<Category> getAllCategorys() {
+    private final CategoryRepository categoryRepository;
+    /**
+     * Retrieve all categories from the database.
+     * @return a list of categories
+     */
+    public List<Category> getAllCategories() {
         return categoryRepository.findAll();
     }
-
-    public Category getCategoryById(Long id) {
-        return categoryRepository.findById(id).orElse(null);
+    /**
+     * Retrieve a category by its id.
+     * @param id the id of the category to retrieve
+     * @return an Optional containing the found category or empty if not found
+     */
+    public Optional<Category> getCategoryById(Long id) {
+        return categoryRepository.findById(id);
+    }
+    /**
+     * Add a new category to the database.
+     * @param category the category to add
+     */
+    public void addCategory(Category category) {
+        categoryRepository.save(category);
+    }
+    /**
+     * Update an existing category.
+     * @param category the category with updated information
+     */
+    public void updateCategory(@NotNull Category category) {
+        Category existingCategory = categoryRepository.findById(category.getId())
+                .orElseThrow(() -> new IllegalStateException("Category with ID " +
+                        category.getId() + " does not exist."));
+        existingCategory.setName(category.getName());
+        categoryRepository.save(existingCategory);
     }
 
-    public void deleteCategory(Long id) {
+
+    public void deleteCategoryById(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new IllegalStateException("Category with ID " + id + " does not * exist.");
+        }
         categoryRepository.deleteById(id);
     }
 
-    public void saveCategory(Category category) {
-        categoryRepository.save(category);
-    }
 }
+
