+++ src/main/java/com/shopthoitrangnike/shopThoiTrangNike/service/ProductService.java
@@ -2,28 +2,48 @@ package com.shopthoitrangnike.shopThoiTrangNike.service;
 
 import com.shopthoitrangnike.shopThoiTrangNike.model.Product;
 import com.shopthoitrangnike.shopThoiTrangNike.repository.ProductRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import jakarta.validation.constraints.NotNull;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
+@Transactional
 public class ProductService {
-    @Autowired
-    private ProductRepository productRepository;
-
-    List<Product> getAllProduct(){ return productRepository.findAll();}
-    public Optional<Product> getProductById(Long id){ return productRepository.findById(id);}
-    public void addProduct(Product product) {
-        productRepository.save(product);
+    private final ProductRepository productRepository;
+    // Retrieve all products from the database
+    public List<Product> getAllProducts() {
+        return productRepository.findAll();
     }
-
-    public void deleteProduct(Long id) {
+    // Retrieve a product by its id
+    public Optional<Product> getProductById(Long id) {
+        return productRepository.findById(id);
+    }
+    // Add a new product to the database
+    public Product addProduct(Product product) {
+        return productRepository.save(product);
+    }
+    // Update an existing product
+    public Product updateProduct(@NotNull Product product) {
+        Product existingProduct = productRepository.findById(product.getId())
+                .orElseThrow(() -> new IllegalStateException("Product with ID " +
+                        product.getId() + " does not exist."));
+        existingProduct.setName(product.getName());
+        existingProduct.setPrice(product.getPrice());
+        existingProduct.setSize(product.getSize());
+        existingProduct.setDescription(product.getDescription());
+        existingProduct.setCategory(product.getCategory());
+        return productRepository.save(existingProduct);
+    }
+    // Delete a product by its id
+    public void deleteProductById(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new IllegalStateException("Product with ID " + id + " does not exist.");
+        }
         productRepository.deleteById(id);
-    }
-
-    public void updateProduct(Product product) {
-        productRepository.save(product);
     }
 }
